source: music: switched to aac audio
finally I got it to playback aac encoded audio.
- removed the audio as wav files
- re-added them as aac compressed audio again in an m4a container

Using the JAAD decoder to playback them, as I didn't want to add all
those lines of code anywhere I'd need them maybe in the future, I
wrote a small library 'AACPlayer', which uses the JAAD decoder.

Therefore, you have to add to the libraries of the project from now on
- lanterna 2.1.9
- JAAD 0.8.4
- AACPlayer 1.0.0